Fixes #4.
This is just a quick-fix however. The mechanism used
to check if an editor should be observed is very suboptimal
currently. For instance a new file (editor) that is later
saved as an ocaml file will not be picked up and linted.

I guess things will also break if a file is moved!

I need to listen for onDidChangePath and add/remove the
editor as need by I suppose.